mcp_client/pyproject.toml
[project]
name = "kuberai-mcp-client"
version = "0.1.0"
description = "MCP Client component for KuberAI"
authors = [
    {name = "Shreyas", email = "shreyasbhat17@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

# Component-specific dependencies
dependencies = [
    "asyncio-mqtt>=0.16.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-mcp-adapters>=0.1.8",
    "fastmcp>=2.10.1",
    "fastapi>=0.115.14"
]

# Optional dependencies for this component
[dependency-groups]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Package discovery
[tool.hatch.build.targets.wheel]
packages = ["src/"]

